test: communication
