+++ client/web/src/app/minihome/[user-name]/guestbook/components/guestbook-content/index.css.ts
@@ -1,0 +1,9 @@
+import { style } from "@vanilla-extract/css";
+
+export const guestbook = style({
+  display: "flex",
+  flexDirection: "column",
+  gap: "20px",
+  paddingTop: "32px",
+  paddingBottom: "32px",
+});
