+++ client/web/src/app/minihome/[user-name]/guestbook/page.css.ts
@@ -3,11 +3,3 @@ import { style } from "@vanilla-extract/css";
 export const h1 = style({
   marginBottom: "48px",
 });
-
-export const guestbook = style({
-  display: "flex",
-  flexDirection: "column",
-  gap: "20px",
-  paddingTop: "32px",
-  paddingBottom: "32px",
-});
