+++ client/web/src/app/minihome/[user-name]/guestbook/page.tsx
@@ -1,20 +1,12 @@
-import { Divider } from "@/components/divider";
-
-import { UserInput } from "./components/user-input";
-import { guestbook, h1 } from "./page.css";
+import { GuestbookContent } from "./components/guestbook-content";
+import { h1 } from "./page.css";
 
 export default function Guestbook() {
   return (
     <>
       <h1 className={h1}>Guestbook</h1>
       <section>
-        <UserInput />
-        <article className={guestbook}>mapping</article>
-        <Divider />
-        <article className={guestbook}>mapping</article>
-        <Divider />
-        <article className={guestbook}>mapping</article>
-        <Divider />
+        <GuestbookContent />
       </section>
     </>
   );
